shell: Add subdir-objects to AM_INIT_AUTOMAKE
I've attached a patch which adds the subdir-objects option to AM_INIT_AUTOMAKE.

For a while now when I've compiled dash I received a warning from
automake that there are source files in a subdirectory but that the
subdir-objects automake option was not supplied. I've just been adding
it myself, but I finally got around to submitting a patch. The code
still compiles for now (i'm using automake 1.15.1), but warning text
is rarely nice to see and, if the warning text is to be believed, then
the warning will eventually become an error.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>