parser: use pgetc_eatbnl() in more places.
gwsh has a pgetc_eatbnl function in parser.c which skips any backslash-newline
combinations. It's not used everywhere it could be. There is also some
duplicated backslash-newline handling elsewhere in parser.c. Replace most of
the calls to pgetc() with calls to pgetc_eatbnl() and remove the duplicated
backslash-newline handling.