Re-work pmatch().
The old version of pmatch() has worst-case exponential complexity.

Test case:

  case xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx in
  x*x*x*x*x*x*x*x*x*x*x*x*x*x*x*x*x*x*x*x*x*x*x*x*x*x*x*x*x*x*x*x*x) ;;
  *) echo ok ;;
  esac

In addition, this new version is capable of doing what scanleft() and
scanright() did. It takes a strict left-to-right scan of the string which
is subobtimal for some cases, but still an improvement overall.

As fnmatch() does not provide those same options, fnmatch() can no longer
be used.

This version uses the fast that a backtracking approach only needs to
consider the last asterisk, as seen in Jilles Tjoelker's implementation in
FreeBSD sh.