shell-quote xtrace output
The xtrace (set -x) output in gwsh is a bit of a pain, because
arguments containing whitespace aren't quoted. This can it
impossible to tell which bit belongs to which argument:

$ gwsh -c 'set -x; printf "%s\n" one "two three" four'
+ printf %s\n one two three four
one
two three
four

Another disadvantage is you cannot simply copy and paste the
commands from this xtrace output to repeat them for debugging
purposes.

This commit introduces shell-quoting of xtrace arguments that
contain non-shell-safe characters or are identical to reserved
words.

A parameter is added to single_quote() indicating whether quoting
should be unconditional (0) or conditional upon the string
containing characters that are not shell-safe (1). This leaves the
unconditional quoting in the output of commands like 'trap' and
'set' unchanged while avoiding excessive quoting in xtrace output.

Shell-safe characters are defined as this set:
0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz%+,./:@_^-

Quoting of assignments needs to be handled specially; in order for
the output to be suitable for re-entry into the shell, only the
part after the "=" must be quoted. For this purpose, eprintlist()
was split in into two functions, eprintvarlist() to print the
variable assignments and eprintarglist() to print the rest.