Use locale during parsing.
When --enable-locale is used, the parser will take locale into account when
determining which bytes form characters and which characters are blanks.
The locale will also be used for $'...' strings and shell quoting of various
command and xtrace output.

When --enable-parser-locale is used, this locale will be set at shell
startup and will not reflect LC_CTYPE changes from within the shell session.
This is required by POSIX.

When --disable-parser-locale is used, the parser will use whatever locale is
the current locale at time of parsing. This matches what most other shells
do.