Avoid undefined behaviour in arithmetic expressions.
Overflow in arithmetic expressions in shell scripts resulted in overflow in
the shell, triggering UBSAN warnings and in some cases results depending on
the target CPU.

Now, all arithmetic is handled consistently. Overflow is reduced modulo
UINTMAX_T+1, shifts of negative values are guaranteed to be evaluated as an
arithmetic shift, shift counts are interpreted as unsigned, and excessive
shift counts do not lose the high bits.