Implement privileged mode.
Originally reported by Tavis Ormandy to dash in 2013, with Jilles Tjoelker
pointing out a possible remaining security issue and a bash
incompatibility. Re-implemented to take that into account and to simplify
the implementation.

By default, when the shell is launched in a privileged context (one where
real UID/GID and effective UID/GID do not match), it launches in privileged
mode. In privileged mode, the start-up files under user control
($HOME/.profile and the file referenced by the ENV variable) are not read,
but otherwise the shell behaves as normal. Privileged mode may be turned
off with 'set +p', which drops any extra permissions.

When configured with --disable-default-privileged, if the shell is launched
in a privileged context, and privileged mode is not specifically requested
with the -p option, it immediately drops any extra permissions. This is not
what POSIX specifies, but more secure.