Let scripts execute with gwsh.
When an executable script without #! is launched, gwsh would invoke it
using _PATH_BSHELL. If possible, launch it using the same gwsh
executable that is currently running, using /proc/self/exe. Add a
configure option to override this.

Also make sure that we do not misrepresent files as options.

Test case:

  echo 'echo ok' > -s
  chmod +x ./-s
  PATH=:$PATH
  -s <<EOF
  echo bad
  EOF