Fix handling of 'command local'.
Because 'command' undoes the special built-in effect of letting
pre-command variable assignments persist after the command finishes,
those variable assignments would be seen as the current scope for
'local'.

This also allows memory to be saved when 'local' is used multiple times
in the same function for the same variable, for instance at the start of
a loop. We do not need to preserve the prior value every time, we only
need to do that the very first time.

Test cases:

  f() {
    a=1 b=1
    a=2 command local a=3 b=2
    echo a=$a b=$b  # a=1 b=2
  }
  f
  command local a=4 # local: not in a function

Seen in FreeBSD sh testsuite.