Fixes for signal handling.
This fixes two separate problems. The first is that the resetting of
signal handlers in subshells was not working at all. The second is that
if a signal was received while traps were being reset, whether because
of a subshell or because of an explicit reset by a trap command, the
signal could be caught and ignored.

Test case:

  trap : TERM
  while :; do :; done &
  kill $!
  wait