Remove %builtin and %func PATH handling.
The use of %builtin and %func in PATH is non-standard and using it poses
a potential security risk if external commands are invoked, as they will
treat such path components as path names and check directories that were
not meant to be checked.

For %builtin, no replacement is provided. For %func, the FPATH
environment variable can be used instead. This provides a subset of the
functionality supported by ksh.

Test case:

  echo 'x() { :; }' >x
  unset FPATH
  x # x: not found
  FPATH=.
  x