Fix use of uninitialized value.
lvp->local needs to be set whenever name is not "-", as it gets used in
poplocalvars(). Without setting it, it remains uninitialized.

vp->local should be set whenever name is not "-", as it gets used to
determine whether a local variable of that name already exists in the
current scope. Without setting it, it keeps its value of NULL.