Change the way variables are stored.
Before this change, variables were stored in memory as either varname\0\0
if the variable was unset, or varname=varvalue\0 is the variable was set
to varvalue. After this change, variables are stored in memory as either
varname\0\0\1 if the variable is unset, or varname\0varvalue\0\0 if the
variable is set to varvalue. This avoids the need to copy variable names
just to compare them according to the current locale, reduces code size
and improves performance of larger scripts, even if it potentially
increases memory usage and shell startup time by requiring environment
variables to be copied.

No change in behaviour is intended.