Fix trap handling for signals that were ignored on entry.
POSIX requires the trap command to not install handlers for signals that
were ignored on entry to non-interactive shells. Previously,
job-control-enabled shells, which need not be interactive, would allow
handlers to be installed for TSTP/TTIN/TTOU. They are now treated the
same as all other signals.

POSIX does allow interactive shells to install handlers for signals that
were ignored on entry. This is something some shells do for some
signals, but ash historically did not treat interactive and
non-interactive modes differently in this regard. For now, that remains
the case.

Test case:

  $SHELL -c '
    trap "" TTIN
    $SHELL -mc "
      trap \"echo TTIN\" TTIN
      kill -TTIN \$\$
    "
  '