Fix set -o handling.
set -o is supposed to accept options as separate words as well as
options that immediately follow the 'o', even though scripts are not
supposed to leave out the space.

It is not possible to support both that and the odd but historically
accepted 'set -of monitor' to turn on both set -f and set -o monitor, so
that now fails. 'set -fo monitor' is still supported.

Test case:

  set -onoglob
  echo $-