Be more strict in keyword recognition.
In for and case commands, we are expected to only recognise a subset of
keywords. Previously, we recognised all keywords but tried to handle
everything that was not in that subset in the ordinary sense. This
approach fails to handle aliases, which are only recognised when the
word is not a keyword.

Test case:

  alias x="case for in _) ;; " for=_
  eval "x for) echo bug ;; for) echo ok ;; esac"