When reading from stdin, do not read ahead.
This required fixing a long-standing bug in the error recovery, where
the discarding of the rest of the line after encountering a syntax error
was done by modifying the input buffer. This did not take into account
the possibility that the full line was not yet read into the input
buffer.

Fixing that, in turn, uncovered an error in handling of invalid
redirections: these are runtime errors, not syntax errors, and reporting
them using synerror reported them with the wrong line number.

Test cases:

  $SHELL <<EOF
  echo hello
  sed -u q
  world
  echo !
  EOF

  ENV= $SHELL -i <<EOF
  )$(yes "" | head -n 10000 | tr "\n" " ")echo bug
  echo ok
  EOF

  $SHELL <<\EOF
  echo >&$PWD
  EOF