Fix handling of invalid substitutions.
Invalid substitutions cannot be encoded using null bytes, as that
prevents duplication of nodes in copynode(). Encode them as VSNUL
instead.

There is also a bug specific to gwsh in the processing of invalid
substitutions: we did not properly skip over invalid substitutions.

The parsing of substitutions is somewhat simplified, and the invalid
${#+} is now no longer taken as ${#+""}, but as an invalid substitution.
It is already treated as an invalid substitution in bash, dash, and ksh.

Based on a report to the dash mailing list by Christoph Anton
Mitterer <calestyo@scientia.org>.

Test case:

  var=Hello $SHELL -c 'f() { echo ${var-${var!}}; }; f'

  $SHELL -c 'echo ${#+}' # Bad substitution