Treat special built-ins non-specially in interactive shells
We implemented the required behaviour of having interactive shells not
exit on shell errors by raising errors all the way back to main() and
handling it there. This is incorrect: this is starting the exit of the
shell and then cancelling it, when what we should be doing is not exit
the shell in the first place. As reported in Debian bug #588111, it
interacts badly with set -e, but it interacts badly with anything else
that checks its exit status as well.

Test case:

  ENV= $SHELL -i <<\EOF
  exec 2>/dev/null
  export -X && echo bug 1
               echo ok  2
  export -X || echo ok  3
  set -e
  export -X || echo ok  4
  export -X && echo bug 5
               echo ok  6
  export -X
               echo bug 7
  EOF