Fix error handling of unterminated command substitutions.
Commit d2795230 changed the parser to skip the rest of the line upon
encountering a syntax error. This also triggers when the lexer calls
synexpect() for an unterminated old-style command substitution, where
the attempt to skip the rest of the line causes unbounded recursion.
Change the state before reporting the error so that we no longer show up
as inside a command substitution.

Test case:

  echo \` | $SHELL # should error, not segfault.