alias: avoid out of bounds access.
If we run alias "", we check whether the argument is a name or an
assignment by looking for the first '\0' or '=' starting from the second
character. If the first character is already '\0', there is no second
character to start from.

This is probably not causing problems, as we prepare command arguments
for commands using expandarg(), which first writes the argument with
escapes, and then removes the escapes in-place, ensuring that after
the initial null byte, we still have CTLESC followed by another null
byte, meaning we end up doing the right thing for the wrong reason.