Simplify alias storage.
Rather than storing the alias name and value separately, we can reduce
simplify code and reduce code size by storing them in name=value form.
This allows us to re-use some code from var.c to handle hashing and
comparisons, so long as we update that to account for aliases' special
handling of a leading = character. This is okay to do for variables as
well, as for variables the leading character is guaranteed to not be =.

This has the benefit of automatically quoting alias names, reported to
the dash mailing list by наб <nabijaczleweli@nabijaczleweli.xyz>.

No other change in behaviour is intended.

Test case:

  alias '*='
  alias '*'