Do not skip line when we are at the start
The change to skip the rest of the line of input on error missed that if
we are at the start of a line, if a newline resulted in a syntax error,
we must not skip another line of input.

Based on a message to the dash mailing list by Herbert Xu.

Test case (in an interactive shell):

  >