mklocal(): avoid use-after-free.
In mklocal(), if we make a variable local that was already made local in
the current scope, we try to avoid creating a new localvar as it leads
to unnecessary increasing memory use if local is used repeatedly in a
loop. Previously, we would pre-emptively create the localvar, and then
free it during the same mklocal() if it turned out we did not need it.
There was a bug here: when we detected that we did not need it, we would
not free only the new localvar, we would also free the existing var's
text which was still in use.

This commit fixes that, and does some extra work to avoid allocating a
new localvar in the first place that we are just going to end up
freeing.

Test case:

  f() {
    local foo=bar foo
  }
  f