Handle completion slightly better when names are not valid characters.
When a file name contains bytes that are not valid in the current
locale, str_width() will return SIZE_MAX. When we then divide by
maxwidth + 1, we would try to divide by zero.

In general, for the purpose of calculating the column width, if we see
any file name that is larger than the screen width, we know we can only
display a single column, so just do that.