Add CMake build support.
This commit adds CMake build support as an alternative to the tried and
tested autoconf/automake build system. The old build system is not at
risk of being removed, but is missing some desirable features that
alternatives do provide, such as the ability to build with Ninja, and
better integration with IDEs.

While looking at the build system, this commit also removes an
unnecessary check for yacc.